+++ src/utils/graphql.js
@@ -35,6 +35,17 @@ const client = new ApolloClient({
 const GET_FORKS_QUERY = gql`
   query Forks($owner: String! $name: String!) {
     repository(owner: $owner, name: $name) {
+      nameWithOwner
+      stargazerCount
+      forkCount
+      object(expression: "master") {
+        ... on Commit {
+          committedDate
+          history {
+            totalCount
+          }
+        }
+      }
       forks(first: 100, orderBy: {field: STARGAZERS, direction: DESC}) {
         nodes {
           nameWithOwner