+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [Unreleased]
+  - Include origin to node list, refs #11
+
 ## [2020.12.19]
   - Sortable columns, refs #1
   - Last commit date bug fix, refs #9